+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Update Font Awesome icons
 
 ## [0.1.6] - 2019-02-22
 ### Added