+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.1.7] - 2020-05-09
 - RuuviTags can be renamed in discovery list
 - Update Font Awesome icons
 - Add a close button to modal
@@ -40,7 +42,8 @@
 
 First release
 
-[Unreleased]: https://github.com/balda/ruuvitag-discovery/compare/0.1.6...HEAD
+[Unreleased]: https://github.com/balda/ruuvitag-discovery/compare/0.1.7...HEAD
+[0.1.7]: https://github.com/balda/ruuvitag-discovery/compare/0.1.6...0.1.7
 [0.1.6]: https://github.com/balda/ruuvitag-discovery/compare/0.1.5...0.1.6
 [0.1.5]: https://github.com/balda/ruuvitag-discovery/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/balda/ruuvitag-discovery/compare/0.1.3...0.1.4